tests/stress/requirements.txt
articlemetaapi==1.13.19
pymongo==3.4.0
requests==2.11.1
thriftpy==0.3.1
